+++ changes.md
@@ -13,14 +13,19 @@ These are the changes for the next update[s].
 
 - Hover over ships to see info
 - Better stars (see https://www.babylonjs-playground.com/#MX2Z99#7)
-- Hardpoints (see https://playground.babylonjs.com/#PU4WYI#291)
+- Hardpoints pt. 2 (see https://playground.babylonjs.com/#PU4WYI#291)
 
 ## Changes since Alpha 1.2.1 release
 
 ### Features
+- Hardpoints
 
 ### Bug Fixes
 - Hyperspace
+- Save date display fixed
+- Icon clicks
 
 ### Technical Changes
-- Further seperation of client and core
+- Changed serialization and deserialization to be part of classes
+- Improvements to entity selection
+- Change settings (incl. keybinds) to use the localStorage API